in fold-tokens, added VisitTokens
in array-expr, started work for better variable interpolation.
